requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
google-genai
Werkzeug==2.3.7
pathlib2==2.3.7
python-dotenv==1.0.0
